requirements.txt
boto3
expiringdict
readerwriterlock
fusepy
moto